+++ engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/KafkaAvroSourceFactory.scala
@@ -1,44 +1,46 @@
 package pl.touk.nussknacker.engine.avro
 
 import javax.validation.constraints.NotBlank
+import org.apache.avro.Schema
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.streaming.api.functions.TimestampAssigner
-import org.apache.flink.streaming.connectors.kafka.KafkaDeserializationSchema
 import pl.touk.nussknacker.engine.api.editor.{DualEditor, DualEditorMode, SimpleEditor, SimpleEditorType}
 import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, Source, TestDataGenerator}
 import pl.touk.nussknacker.engine.api.test.TestParsingUtils
 import pl.touk.nussknacker.engine.api.typed.{ReturningType, typing}
 import pl.touk.nussknacker.engine.api.{MetaData, MethodToInvoke, ParamName}
-import pl.touk.nussknacker.engine.avro.formatter.AvroToJsonFormatter
-import pl.touk.nussknacker.engine.avro.typed.AvroSchemaTypeDefinitionExtractor
+import pl.touk.nussknacker.engine.avro.fixed.FixedKafkaAvroSchemaProvider
+import pl.touk.nussknacker.engine.avro.schemaregistry.{SchemaRegistryKafkaAvroProvider, SchemaRegistryProvider}
 import pl.touk.nussknacker.engine.kafka.KafkaSourceFactory._
 import pl.touk.nussknacker.engine.kafka._
-import pl.touk.nussknacker.engine.kafka.serialization.DeserializationSchemaFactory
 
-class KafkaAvroSourceFactory[T: TypeInformation](schemaFactory: DeserializationSchemaFactory[T],
-                                                 schemaRegistryClientFactory: SchemaRegistryClientFactory,
-                                                 timestampAssigner: Option[TimestampAssigner[T]],
-                                                 formatKey: Boolean = false,
-                                                 processObjectDependencies: ProcessObjectDependencies)
-  extends KafkaSourceFactory[T](schemaFactory, timestampAssigner,
-    TestParsingUtils.newLineSplit, processObjectDependencies) {
+class KafkaAvroSourceFactory[T: TypeInformation](schemaRegistryProvider: SchemaRegistryProvider[T],
+                                                 processObjectDependencies: ProcessObjectDependencies,
+                                                 timestampAssigner: Option[TimestampAssigner[T]])
+  extends BaseKafkaAvroSourceFactory[T](processObjectDependencies, timestampAssigner) {
 
-  override protected def createSource(processMetaData: MetaData,
-                                      topics: List[String],
-                                      schema: KafkaDeserializationSchema[T]): KafkaSource = {
-    val schemaRegistryClient = schemaRegistryClientFactory.createSchemaRegistryClient(kafkaConfig)
-    new KafkaSource(consumerGroupId = processMetaData.id, topics, schema,
-      Some(AvroToJsonFormatter(schemaRegistryClient, topics.head, formatKey)), processObjectDependencies)
-  }
-
+  @MethodToInvoke
+  def create(processMetaData: MetaData,
+             @ParamName(`TopicParamName`)
+             @DualEditor(
+               simpleEditor = new SimpleEditor(`type` = SimpleEditorType.STRING_EDITOR),
+               defaultMode = DualEditorMode.RAW
+             )
+             @NotBlank
+             topic: String): Source[T] with TestDataGenerator =
+    createKafkaAvroSource(processMetaData, topic, SchemaRegistryKafkaAvroProvider(schemaRegistryProvider, kafkaConfig, topic))
 }
 
-class KafkaTypedAvroSourceFactory[T: TypeInformation](schemaFactory: DeserializationSchemaFactory[T],
-                                                      schemaRegistryClientFactory: SchemaRegistryClientFactory,
-                                                      timestampAssigner: Option[TimestampAssigner[T]],
-                                                      formatKey: Boolean = false,
-                                                      processObjectDependencies: ProcessObjectDependencies)
-  extends BaseKafkaSourceFactory[T](timestampAssigner, TestParsingUtils.newLineSplit, processObjectDependencies) {
+object KafkaAvroSourceFactory {
+  def apply[T: TypeInformation](schemaRegistryProvider: SchemaRegistryProvider[T], processObjectDependencies: ProcessObjectDependencies): KafkaAvroSourceFactory[T] =
+    new KafkaAvroSourceFactory(schemaRegistryProvider, processObjectDependencies, None)
+}
+
+class FixedKafkaAvroSourceFactory[T: TypeInformation](processObjectDependencies: ProcessObjectDependencies,
+                                                      formatKey: Boolean,
+                                                      useSpecificAvroReader: Boolean,
+                                                      timestampAssigner: Option[TimestampAssigner[T]])
+  extends BaseKafkaAvroSourceFactory[T](processObjectDependencies, timestampAssigner) {
 
   @MethodToInvoke
   def create(processMetaData: MetaData,
@@ -49,15 +51,36 @@ class KafkaTypedAvroSourceFactory[T: TypeInformation](schemaFactory: Deserializa
              )
              @NotBlank
              topic: String,
-             // json or avro schema on this level?
              @ParamName("schema")
              @NotBlank
-             avroSchema: String): Source[T] with TestDataGenerator = {
-    val schemaRegistryClient = schemaRegistryClientFactory.createSchemaRegistryClient(kafkaConfig)
-    new KafkaSource(consumerGroupId = processMetaData.id, List(topic), schemaFactory.create(List(topic), kafkaConfig),
-      Some(AvroToJsonFormatter(schemaRegistryClient, topic, formatKey)), processObjectDependencies) with ReturningType {
-      override def returnType: typing.TypingResult = AvroSchemaTypeDefinitionExtractor.typeDefinition(avroSchema)
-    }
-  }
-
+             @SimpleEditor(`type` = SimpleEditorType.STRING_EDITOR)
+             //TODO: Create BE and FE validator for verify avro type
+             //TODO: Create Avro Editor
+             avroSchema: String): Source[T] with TestDataGenerator =
+    createKafkaAvroSource(
+      processMetaData,
+      topic,
+      new FixedKafkaAvroSchemaProvider(topic, avroSchema, kafkaConfig, formatKey, useSpecificAvroReader)
+    )
 }
+
+object FixedKafkaAvroSourceFactory {
+  def apply[T: TypeInformation](processObjectDependencies: ProcessObjectDependencies): FixedKafkaAvroSourceFactory[T] =
+    new FixedKafkaAvroSourceFactory(processObjectDependencies, formatKey = false, useSpecificAvroReader = false, timestampAssigner = None)
+}
+
+abstract class BaseKafkaAvroSourceFactory[T: TypeInformation](processObjectDependencies: ProcessObjectDependencies, timestampAssigner: Option[TimestampAssigner[T]])
+  extends BaseKafkaSourceFactory(timestampAssigner, TestParsingUtils.newLineSplit, processObjectDependencies) {
+
+  def createKafkaAvroSource(processMetaData: MetaData, topic: String, kafkaAvroSchemaProvider: KafkaAvroSchemaProvider[T]): KafkaSource =
+    new KafkaSource(
+      consumerGroupId = processMetaData.id,
+      List(topic),
+      kafkaAvroSchemaProvider.deserializationSchema,
+      kafkaAvroSchemaProvider.recordFormatter,
+      processObjectDependencies
+    ) with ReturningType {
+      override def returnType: typing.TypingResult = kafkaAvroSchemaProvider.typeDefinition
+    }
+}
+