+++ docs/Changelog.md
@@ -15,6 +15,8 @@ By default setted up to false.
 * Old way of configuring Flink and model (via `flinkConfig` and `processConfig`) is removed. `processTypes` 
 configuration should be used from now on.
 * Change of additional properties configuration
+* [#871](https://github.com/TouK/nussknacker/pull/871) Added SchemaRegistryProvider
+* [#881](https://github.com/TouK/nussknacker/pull/881) Introduction to KafkaAvroSchemaProvider
 
 0.0.12 (26 Oct 2019)
 --------------------